refactor: PostDetailModal 컴포넌트 분리 - 게시물 상세보기
